requirements.txt
accelerate==1.10.1
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.2
controlnet-aux==0.0.3
cycler==0.12.1
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1756742004/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
diffusers==0.27.2
distro==1.9.0
easydict==1.13
einops==0.8.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.108.0
ffmpy==0.6.1
filelock==3.13.1
Flask==2.2.3
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flaskwebgui==0.3.5
fonttools==4.59.2
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
gradio==4.21.0
gradio_client==0.12.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.25.2
idna==3.10
imageio==2.37.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
IOPaint==1.6.0
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
kiwisolver==1.4.9
lama-cleaner==1.2.5
lazy_loader==0.4
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.4.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
openai==1.107.0
opencv-python==4.11.0.86
orjson==3.11.3
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.2
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
peft==0.7.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
piexif==1.1.3
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
polars==1.33.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==6.32.0
psutil @ file:///D:/bld/psutil_1755851332307/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
PyYAML==6.0.2
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
ruff==0.13.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.32.0.post1
streamlit==1.49.1
sympy==1.13.3
tenacity==9.1.2
tifffile==2025.5.10
timm==1.0.19
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.12.0
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tornado @ file:///D:/bld/tornado_1756854837538/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.48.3
typer==0.17.4
typer-config==1.4.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.201
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websockets==11.0.3
Werkzeug==2.2.2
whichcraft==0.6.1
win32_setctime==1.2.0
wsproto==1.2.0
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
